Backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
black==26.1.0
click==8.3.1
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.8.0
drf-yasg==1.21.14
inflection==0.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
pytokens==0.4.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
